# openapi/drouple.v1.yaml
openapi: 3.0.3
info:
  title: Drouple Unified API
  version: 1.0.0
  description: Unified API for both Web and Mobile clients
  contact:
    name: Drouple API Team
    email: api@drouple.com
servers:
  - url: https://api.drouple.com/api/v1
    description: Production server
  - url: https://api-staging.drouple.com/api/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      required: [id, email, name, role, tenantId]
      properties:
        id:
          type: string
          example: "user_123"
        email:
          type: string
          format: email
          example: "user@drouple.com"
        name:
          type: string
          example: "John Doe"
        role:
          type: string
          enum: [SUPER_ADMIN, CHURCH_ADMIN, VIP, LEADER, MEMBER]
          example: "MEMBER"
        tenantId:
          type: string
          example: "tenant_123"
        churchId:
          type: string
          nullable: true
          example: "church_123"
...
security:
  - BearerAuth: []
paths:
...
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
                  example: "refresh_token_here"
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
                  code:
                    type: string
                    example: "TOKEN_REFRESHED"
                  data:
                    type: object
                    properties:
                      accessToken:
                        type: string
                        example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                      refreshToken:
                        type: string
                        example: "new_refresh_token_here"
...
  /members/search:
    get:
      tags: [Members]
      summary: Search members
      parameters:
        - name: q
          in: query
          description: Search query
          schema:
            type: string
            example: "john"
        - name: limit
          in: query
          description: Max results to return
          schema:
            type: integer
            default: 20
            example: 20
        - name: offset
          in: query
          description: Number of results to skip
          schema:
            type: integer
            default: 0
            example: 0
      responses:
        '200':
          description: Members found
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      members:
                        type: array
                        items:
                          $ref: '#/components/schemas/User'
                      total:
                        type: integer
                        example: 50
...